requirements.txt
numpy
pandas
plotly.express
streamlit
streamlit_option_menu